+++ NEWS.md
@@ -8,7 +8,8 @@
 
 ---
 ## (current master)
-- preallocate identity matrix
+- Add interpolateDataFrameOntoδ function. This is used to convert a  measured size distribution stored as a DataFrame and to a SizeDistribution that matches a DMA grid. The function simplifies using the package with real data.
+- Preallocate identity matrix
 - Fix performance regression caused by changes in 2.1
 - Add subtraction operator: 𝕟1 - 𝕟2
 - Remove Plots.jl/PlotlyJS.jl/ORCA.jl dependencies for core project. The packages are still needed for executing the notebooks but they don't need to be installed when the package is embedded for regular use and carry a lot of overhead.